pyproject.toml
[tool.poetry]
name = "stock-pulse"
version = "0.1.0"
description = ""
authors = ["Amit Wani <mtwn105@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.3.25"
langchain-openai = "^0.3.16"
langgraph = "^0.4.3"
yfinance = "^0.2.59"
python-dotenv = "^1.1.0"
rich = "^14.0.0"
pydantic = "^2.11.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
